pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autojudge-base"
version = "0.1.0"
description = "Core infrastructure for implementing TREC AutoJudge systems"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "TREC AutoJudge Team"}
]
keywords = ["trec", "autojudge", "rag", "evaluation", "ir"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "click>=8.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "pandas>=2.0",
    "ir-datasets>=0.5",
    "nltk>=3.8",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]
minima-llm = [
    "minima-llm>=0.1.0",
]

[project.scripts]
auto-judge = "autojudge_base.cli:main"

[project.urls]
Homepage = "https://github.com/trec-autojudge/auto-judge-base"
Repository = "https://github.com/trec-autojudge/auto-judge-base"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]